requirements.txt
pandas==2.2.3
google-api-python-client==2.170.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2